Refactor how creature colors work
Now items and enchantments can return a new creature color
and priority. The final creature color is determined by
the enchantment or item which returned the color with
highest priority.

This means that Saboteurs are always red, regardless of
enchantments, and that miners with a firestarter spell
will always be red too